Base implementation of FileSystemElement
The idea is to apply the element to be rewriten, aka rebase, in a
different directory and give access to the file themselves. However, the
tricky part is to still allow some logical manipulation so we can find
the IncrementalElement and/or apply changes after the fact base on the
logical structure of the element.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>